I think this fixes the densityOrder sometimes getting stuck on "dense".
It was because we were using zero-height <canvas> elements to signify
that there was too much data to draw, when calculating densityOrder.
However, there were also cases in which a <canvas> would have no data to
draw and therefore would be set to zero-height. To fix this, we make
sure any successful canvas draw results in a >=1px high canvas, and it's
documented in densityOrder, each of the custom track types, and the
README there.