add a default action for clicking on custom track items
which opens the item data in a new window. This is based on tipTipData
but specifically includes "hidden" tipTipData fields, which are prefixed
with "." (period). That way you can have a small version of item data
for the tooltip and then a bigger look after you click (with cigar,
sequences, etc.)

Also, we've changed the default font away from Lucida Grande